move SessionManagementException to api.exception package
