+++ components/uuf-core/src/main/java/org/wso2/carbon/uuf/api/auth/InMemorySessionManager.java
@@ -20,8 +20,8 @@ package org.wso2.carbon.uuf.api.auth;
 
 import com.google.common.collect.Iterables;
 import org.wso2.carbon.uuf.api.config.Configuration;
+import org.wso2.carbon.uuf.api.exception.SessionManagementException;
 import org.wso2.carbon.uuf.api.exception.UUFRuntimeException;
-import org.wso2.carbon.uuf.exception.SessionManagementException;
 import org.wso2.carbon.uuf.spi.HttpRequest;
 import org.wso2.carbon.uuf.spi.HttpResponse;
 import org.wso2.carbon.uuf.spi.auth.SessionManager;
